services/syncthing/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
